+++ sigf_v1/src/main/java/com/areatecnica/sigf_v1/controllers/PlanillonImposicionesController.java
@@ -95,7 +95,7 @@ public class PlanillonImposicionesController implements Serializable {
     private static final int VALORDIA = (int) SUELDOBASE / 30;
     private static final long VALORSIS = (long) 0.0141;
     private Date FECHACESANTIA;
-    private int sueldoAjustadoAux;
+    private float sueldoAjustadoAux;
 
     /**
      * Creates a new instance of PlanillonImposicionesController
@@ -651,7 +651,7 @@ public class PlanillonImposicionesController implements Serializable {
         this.idOperador = idOperador;
     }
 
-    public int getSueldoAjustadoAux() {
+    public float getSueldoAjustadoAux() {
         return sueldoAjustadoAux;
     }
